requirements.txt
streamlit
cryptography
Pillow
